Finish 05-flex-modal
This part of the activity is really wild since its
something i haven't seen before with some parts of
the content is overlapping, I tried solutions with
negative margin on it, end up doing lots of divs &
doing display:flex; on it. Oh well I somewhat ach-
ieved it, in a cleverly complex manner lol.